chores: Organizing files
