+++ package.json
@@ -7,7 +7,7 @@
     "test": "test"
   },
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "jest --coverage --verbose"
   },
   "repository": {
     "type": "git",
@@ -31,5 +31,8 @@
     "got": "^11.5.2",
     "marked": "^1.1.1",
     "readdirp": "^3.4.0"
+  },
+  "devDependencies": {
+    "jest": "^26.4.2"
   }
 }
