+++ test/fixtures/example1.md
@@ -1,0 +1,2 @@
+# Example #1 markdown file
+It's very easy to make some words **bold** and other words *italic* with Markdown. You can even [link to Google!](http://google.com)
